Various typecasting fixes, mis-typed enums, etc.
Details:
- Fixed implicit typecasting of conj_t to trans_t in bli_[un]packm_cxk.c.
- Properly typecast integer arguments to match format specifier in various
  calls to printf() in bli_l3_thrinfo.c, bli_cntx.c, bli_pool.c, and
  bli_util_oapi.c.
- Fixed "unsigned less-than-comparison with zero" checks in bli_check.c,
  bli_cntx.h.
- Fixed mis-typed enums in bli_cntx.c (e.g., l1mkr_t that should have been
  l1fkr_t or l1vkr_t).
- Fixed instances of opid_t value BLIS_GEMM that should have been l3ukr_t
  value BLIS_GEMM_UKR in bli_cntx_ref.c.
- NOTE: These issues were identified via compiler warnings when building
  BLIS with clang on a rather old installation of OS X:
    $ clang --version
    Apple LLVM version 5.0 (clang-500.2.79) (based on LLVM 3.3svn)
    Target: x86_64-apple-darwin15.2.0
    Thread model: posix

+++ frame/base/bli_pool.c
@@ -393,16 +393,17 @@ void bli_pool_print( pool_t* pool )
 
 	printf( "pool struct ---------------\n" );
 	printf( "  block_ptrs:      %p\n", block_ptrs );
-	printf( "  block_ptrs_len:  %ld\n", block_ptrs_len );
-	printf( "  top_index:       %ld\n", top_index );
-	printf( "  num_blocks:      %ld\n", num_blocks );
-	printf( "  block_size:      %ld\n", block_size );
-	printf( "  align_size:      %ld\n", align_size );
+	printf( "  block_ptrs_len:  %ld\n", ( long )block_ptrs_len );
+	printf( "  top_index:       %ld\n", ( long )top_index );
+	printf( "  num_blocks:      %ld\n", ( long )num_blocks );
+	printf( "  block_size:      %ld\n", ( long )block_size );
+	printf( "  align_size:      %ld\n", ( long )align_size );
 	printf( "  pblks   sys    align\n" );
 	for ( i = 0; i < num_blocks; ++i )
 	{
-	printf( "  %ld: %p %p\n", i, bli_pblk_buf_sys(   &block_ptrs[i] ),
-	                             bli_pblk_buf_align( &block_ptrs[i] ) );
+		printf( "  %ld: %p %p\n", ( long )i,
+		                          bli_pblk_buf_sys(   &block_ptrs[i] ),
+	                              bli_pblk_buf_align( &block_ptrs[i] ) );
 	}
 }
